+++ frontend/src/components/main/body/Bodymain.js
@@ -3,7 +3,7 @@ import logo from "../../../img/Designer.jpeg"
 import { Slide } from 'react-slideshow-image';
 import 'react-slideshow-image/dist/styles.css'
 import { useState, useEffect } from 'react'
-import { Link } from "react-router-dom"
+import { useNavigate, Link } from "react-router-dom"
 
 const spanStyle = {
     padding: '20px',
@@ -38,6 +38,7 @@ const Bodymain = () => {
 	const [yourBest,setYourBest] = useState([])
 	const [category,setCategory] = useState('')
 	const [categoryBest,setCategoryBest] = useState([])
+	const navigate = useNavigate()
 
     const myStyle = {
         aspectRatio: '1000/200',
@@ -76,21 +77,10 @@ const Bodymain = () => {
 		setYourBest(yourBest_)
 	}
 	
-	const getCategoryBest = async () => {
-		if(category){
-			const res = await fetch(`http://localhost:4000/payment/categoryBest?category=${category}`)
-			const categoryBest_ = await res.json()
-			setCategoryBest(categoryBest_)
-		}
-	}
 
 	useEffect(() => {
 		getBest()
 	},[])
-	
-	useEffect(() => {
-		getCategoryBest()
-	},[category])
 
     return (
         <div className="p-6 max-w-screen-xl px-40 mx-40" style={{textAlign :"center" , margin:"0px auto"}}>
@@ -107,66 +97,26 @@ const Bodymain = () => {
                 <h2 className="text-xl font-bold mb-4">카테고리별 클래스 모음</h2>
                 <div className="flex flex-col sm:flex-row  hide-scrollbar mb-4" >
                     <div className="flex flex-col items-center mr-8">
-                        <img onClick={() => setCategory('Posture')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%EC%9E%90%EC%84%B8%EA%B5%90%EC%A0%95.png" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
+                        <img onClick={() => navigate('category?category=Posture')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%EC%9E%90%EC%84%B8%EA%B5%90%EC%A0%95.png" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
                         <span className="text-lg">자세개선</span>
                     </div>
                     <div className="flex flex-col items-center mr-8">
-                        <img onClick={() => setCategory('Yoga')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%EC%9A%94%EA%B0%80.jpg" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
+                        <img onClick={() => navigate('category?category=Yoga')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%EC%9A%94%EA%B0%80.jpg" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
                         <span className="text-lg">요가</span>
                     </div>
                     <div className="flex flex-col items-center mr-8">
-                        <img onClick={() => setCategory('Pilates')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%ED%95%84%EB%9D%BC%ED%85%8C%EC%8A%A4.png" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
+                        <img onClick={() => navigate('category?category=Pilates')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%ED%95%84%EB%9D%BC%ED%85%8C%EC%8A%A4.png" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
                         <span className="text-lg">필라테스</span>
                     </div>
                     <div className="flex flex-col items-center mr-8">
-                        <img onClick={() => setCategory('Fitness')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%ED%97%AC%EC%8A%A4.png" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
+                        <img onClick={() => navigate('category?category=Fitness')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%ED%97%AC%EC%8A%A4.png" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
                         <span className="text-lg">헬스</span>
                     </div>
                     <div className="flex flex-col items-center mr-8">
-                        <img onClick={() => setCategory('Ballet')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%EB%B0%9C%EB%A0%88.png" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
+                        <img onClick={() => navigate('category?category=Ballet')} src="https://chatpt-githubaction-s3-bucket.s3.ap-northeast-2.amazonaws.com/images/%EB%B0%9C%EB%A0%88.png" className="relative flex shrink-0 overflow-hidden w-24 h-24 border rounded-full mb-2"></img>
                         <span className="text-lg">발레</span>
                     </div>
                 </div>
-				<div className="max-w-screen-xl mx-auto px-4">
-                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-5 xl:grid-cols-5 gap-8">
-				{categoryBest.length>0 && categoryBest.map(product => (
-                <Link to = {`product/${product.product_id}`}><div key ={product.product_id} className="rounded-lg overflow-hidden">
-                    <img
-                        src={product.product_thumbnail}
-                        alt="Course thumbnail"
-                        className="w-full h-36 object-cover"
-                        width="240"
-                        height="160"
-                        style={{ aspectratio: 240 / 160, objectfit: "cover" }}
-                    />
-                    <div className="p-4">
-                        <h3 className="text-lg font-semibold mb-2">{product.product_name}</h3>
-                        <p className="text-sm mb-4">{product.product_intro}</p>
-                        <div className="flex items-center justify-between mb-2">
-                            <div className="flex items-center" style={{marginLeft :"80%"}}>
-                                <svg
-                                    xmlns="http://www.w3.org/2000/svg"
-                                    width="24"
-                                    height="24"
-                                    viewBox="0 0 24 24"
-                                    fill="none"
-                                    stroke="currentColor"
-                                    strokeWidth="2"
-                                    strokeLinecap="round"
-                                    strokeLinejoin="round"
-                                    className="text-yellow-400 w-4 h-4"
-                                >
-                                    <polygon points="12 2 15.09 8.26 22 9.27 17 14.14 18.18 21.02 12 17.77 5.82 21.02 7 14.14 2 9.27 8.91 8.26 12 2"></polygon>
-                                </svg>
-                                <span className="text-xs font-semibold ml-1">4.5</span>
-                            </div>
-                        </div>
-                    </div>
-                </div></Link>
-
-
-            ))}
-            </div></div>
             </section>
             <h2 className="text-xl font-bold mb-4">당신에게 추천하는 강의</h2>
                 <div className="max-w-screen-xl mx-auto px-4">
